feat(clamav): Add Deployment alternative
Currently, ClamAV is always deployed as StatefulSet.

This doesn't make sense when not using a PersistentVolume (which isn't even the deafult) and has the disadvantage of less flexible rollingUpdate.

This commit adds the possibility to use a Deployment instead of a StatefulSet when the persistentVolume is not enabled.

It also adds the possibility to set updateStrategy options via Values.